backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
gunicorn==20.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
pymongo==4.6.1
cloudinary==1.36.0
dnspython==2.4.2
